Move unit tests to the `tests/unit/` tree.
Also move the pydecred/dcrdata network tests to `tests/integration/`.
