+++ tests/unit/crypto/test_bytearray.py
