+++ tests/unit/crypto/test_mnemonic.py
