+++ tests/unit/pydecred/test-data/sighash.json
